Fix bug calculator cost.

+++ src/main/java/com/cyan/hotel/dto/OptionDTO.java
@@ -1,10 +1,10 @@
 package com.cyan.hotel.dto;
 
 public class OptionDTO {
-  private Float optionCost;
+  private Long optionCost;
   private String optionName;
 
-  public OptionDTO(Float optionCost, String optionName) {
+  public OptionDTO(Long optionCost, String optionName) {
     this.optionCost = optionCost;
     this.optionName = optionName;
   }
@@ -13,11 +13,11 @@ public class OptionDTO {
 
   }
 
-  public Float getOptionCost() {
+  public Long getOptionCost() {
     return optionCost;
   }
 
-  public void setOptionCost(Float optionCost) {
+  public void setOptionCost(Long optionCost) {
     this.optionCost = optionCost;
   }
